wip: add `Sentinel` to `_util`
