+++ src/awkward/_broadcasting.py
@@ -20,7 +20,7 @@ from awkward._parameters import (
     parameters_intersect,
 )
 from awkward._typing import Any, Callable, Dict, List, TypeAlias, Union
-from awkward._util import unset
+from awkward._util import Sentinel, unset
 from awkward.contents.bitmaskedarray import BitMaskedArray
 from awkward.contents.bytemaskedarray import ByteMaskedArray
 from awkward.contents.content import Content
@@ -173,21 +173,6 @@ def all_same_offsets(backend: Backend, inputs: list) -> bool:
             return False
 
     return True
-
-
-# TODO: move to _util or another module
-class Sentinel:
-    """A class for implementing sentinel types"""
-
-    def __init__(self, name, module=None):
-        self._name = name
-        self._module = module
-
-    def __repr__(self):
-        if self._module is not None:
-            return f"{self._module}.{self._name}"
-        else:
-            return f"{self._name}"
 
 
 NO_PARAMETERS = Sentinel("NO_PARAMETERS", __name__)
